hosts/common - gate nvd if only one generation

+++ hosts/common/nix.nix
@@ -60,12 +60,14 @@
   system = {
     activationScripts.report-changes = ''
       PATH=$PATH:${lib.makeBinPath [ pkgs.nvd pkgs.nix ]}
-      nvd diff $(ls -dv /nix/var/nix/profiles/system-*-link | tail -2)
-      mkdir -p /var/log/activations
-      _nvddate=$(date +'%Y%m%d%H%M%S')
-      nvd diff $(ls -dv /nix/var/nix/profiles/system-*-link | tail -2) > /var/log/activations/$_nvddate-$(ls -dv /nix/var/nix/profiles/system-*-link | tail -1 | cut -d '-' -f 2)-$(readlink $(ls -dv /nix/var/nix/profiles/system-*-link | tail -1) | cut -d - -f 4-).log
-      if grep -q "No version or selection state changes" "/var/log/activations/$_nvddate-$(ls -dv /nix/var/nix/profiles/system-*-link | tail -1 | cut -d '-' -f 2)-$(readlink $(ls -dv /nix/var/nix/profiles/system-*-link | tail -1) | cut -d - -f 4-).log" ; then
-        rm -rf "/var/log/activations/$_nvddate-$(ls -dv /nix/var/nix/profiles/system-*-link | tail -1 | cut -d '-' -f 2)-$(readlink $(ls -dv /nix/var/nix/profiles/system-*-link | tail -1) | cut -d - -f 4-).log"
+      if [ $(ls -dv /nix/var/nix/profiles/system-*-link 2>/dev/null | wc -l) -ge 2 ]; then
+        nvd diff $(ls -dv /nix/var/nix/profiles/system-*-link | tail -2)
+        mkdir -p /var/log/activations
+        _nvddate=$(date +'%Y%m%d%H%M%S')
+        nvd diff $(ls -dv /nix/var/nix/profiles/system-*-link | tail -2) > /var/log/activations/$_nvddate-$(ls -dv /nix/var/nix/profiles/system-*-link | tail -1 | cut -d '-' -f 2)-$(readlink $(ls -dv /nix/var/nix/profiles/system-*-link | tail -1) | cut -d - -f 4-).log
+        if grep -q "No version or selection state changes" "/var/log/activations/$_nvddate-$(ls -dv /nix/var/nix/profiles/system-*-link | tail -1 | cut -d '-' -f 2)-$(readlink $(ls -dv /nix/var/nix/profiles/system-*-link | tail -1) | cut -d - -f 4-).log" ; then
+          rm -rf "/var/log/activations/$_nvddate-$(ls -dv /nix/var/nix/profiles/system-*-link | tail -1 | cut -d '-' -f 2)-$(readlink $(ls -dv /nix/var/nix/profiles/system-*-link | tail -1) | cut -d - -f 4-).log"
+        fi
       fi
     '';
